Adapt bubble state nav projector (Work in progress, not working)
